requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
djangorestframework==3.7.7
et-xmlfile==1.0.1
idna==2.6
jdcal==1.3
openpyxl==2.4.9
pytz==2017.3
requests==2.18.4
urllib3==1.22
